+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## v 0.6.3
 
+* grammar for new source `Datasets`
+
+## v 0.6.3
+
 * Refactor command line options. Try it with `dimcli -h`
 * New command for resolving Dimensions IDs into URLs `dimcli -id [value]`
 * Grammar for DSL V 1.22